renal system edits
I updated instances and annotations in the renal system terms. Note: I
cannot remove the OCLC id from the definition field for the “flame
cell” term - I wanted to move that to a database xref field, but I
couldn’t access it in the @ field.  We are also missing a reference for
the “flame bulb” term.