dumping terms in parent classes
moved all YES binned terms related to Cell, Nervous System and
Immaterial Anatomical Entity into the appropriate parent bin. These do
not have IDS and have been given “NEEDS ID etc” as a suffix in the in
label annotation- so that I can organize each parent and work through
the greater organization of a section prior to the tedium of ID and
annotation and relationships.